+++ frontends/web/src/components/rates/rates.tsx
@@ -51,6 +51,7 @@ export const currenciesWithDisplayName: FiatWithDisplayName[] = [
   { currency: 'ILS', displayName: 'Israeli New Shekel' },
   { currency: 'JPY', displayName: 'Japanese Yen' },
   { currency: 'KRW', displayName: 'South Korean Won' },
+  { currency: 'NGN', displayName: 'Nigerian Naira' },
   { currency: 'NOK', displayName: 'Norwegian Krone' },
   { currency: 'PLN', displayName: 'Polish Zloty' },
   { currency: 'RUB', displayName: 'Russian ruble' },
@@ -59,7 +60,7 @@ export const currenciesWithDisplayName: FiatWithDisplayName[] = [
   { currency: 'USD', displayName: 'United States Dollar' },
   { currency: 'BTC', displayName: 'Bitcoin' }
 ];
-export const currencies: Fiat[] = ['AUD', 'BRL', 'CAD', 'CHF', 'CNY', 'CZK', 'EUR', 'GBP', 'HKD', 'ILS', 'JPY', 'KRW', 'NOK', 'PLN', 'RUB', 'SEK', 'SGD', 'USD', 'BTC'];
+export const currencies: Fiat[] = ['AUD', 'BRL', 'CAD', 'CHF', 'CNY', 'CZK', 'EUR', 'GBP', 'HKD', 'ILS', 'JPY', 'KRW', 'NGN', 'NOK', 'PLN', 'RUB', 'SEK', 'SGD', 'USD', 'BTC'];
 
 export const store = new Store<SharedProps>({
   active: 'USD',